+++ Samples/DITranquillityTest/DITranquillityTest/DITranquillityTests_Storyboard.swift
@@ -24,16 +24,17 @@ class DITranquillityTests_Storyboard: XCTestCase {
   func test01_InitialViewController() {
     let builder = DIContainerBuilder()
     
-    builder.register(FooService.self)
-      .asType(ServiceProtocol.self)
-      .initializer { FooService() }
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial(FooService.init)
     
 		builder.register(vc: TestViewController.self)
-      .dependency { (scope, vc) in vc.service = *!scope }
+      .injection{ $0.service = $1 }
     
     let storyboard = try! DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: builder.build())
 
-    let viewController = storyboard.instantiateInitialViewController()
+		let navigation: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+		let viewController = navigation.childViewControllers[0]
     XCTAssert(viewController is TestViewController)
     guard let testVC = viewController as? TestViewController else {
       XCTFail("incorrect View Controller")
@@ -45,12 +46,12 @@ class DITranquillityTests_Storyboard: XCTestCase {
   func test02_ViewControllerByIdentity() {
     let builder = DIContainerBuilder()
     
-    builder.register(FooService.self)
-      .asType(ServiceProtocol.self)
-      .initializer { FooService() }
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial(FooService.init)
     
 		builder.register(vc: TestViewController2.self)
-      .dependency { (scope, vc) in vc.service = *!scope }
+      .injection{ scope, vc in try vc.service = *scope }
     
     let storyboard = try! DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: builder.build())
     
@@ -76,19 +77,20 @@ class DITranquillityTests_Storyboard: XCTestCase {
   func test04_ViewControllerSequence() {
     let builder = DIContainerBuilder()
     
-    builder.register(FooService.self)
-      .asType(ServiceProtocol.self)
-      .initializer { FooService() }
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial{ FooService() }
     
 		builder.register(vc: TestViewController.self)
-      .dependency { (scope, vc) in vc.service = *!scope }
+      .injection { vc, service in vc.service = service }
     
 		builder.register(vc: TestViewController2.self)
-      .dependency { (scope, vc) in vc.service = *!scope }
+      .injection { scope, vc in try vc.service = *scope }
 
     let storyboard = try! DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: builder.build())
     
-    let viewController = storyboard.instantiateInitialViewController()
+		let navigation: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+		let viewController = navigation.childViewControllers[0]
     XCTAssert(viewController is TestViewController)
     guard let testVC = viewController as? TestViewController else {
       XCTFail("incorrect View Controller")
@@ -96,7 +98,7 @@ class DITranquillityTests_Storyboard: XCTestCase {
     }
     XCTAssertEqual(testVC.service.foo(), "foo")
     
-    viewController?.performSegue(withIdentifier: "ShowTestViewController2", sender: nil)
+    viewController.performSegue(withIdentifier: "ShowTestViewController2", sender: nil)
     
     //...
   }
@@ -104,12 +106,12 @@ class DITranquillityTests_Storyboard: XCTestCase {
   func test05_ViewControllerShortRegister() {
     let builder = DIContainerBuilder()
     
-    builder.register(FooService.self)
-      .asType(ServiceProtocol.self)
-      .initializer { FooService() }
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial(FooService.init)
     
     builder.register(vc: TestViewController2.self)
-      .dependency { (scope, vc) in vc.service = *!scope }
+      .injection { scope, vc in try vc.service = *scope }
     
     let storyboard = try! DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: builder.build())
     
@@ -126,21 +128,22 @@ class DITranquillityTests_Storyboard: XCTestCase {
   func test06_UIStoryboard() {
     let builder = DIContainerBuilder()
     
-    builder.register(FooService.self)
-      .asType(ServiceProtocol.self)
-      .initializer { FooService() }
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial{ FooService() }
     
 		builder.register(vc: TestViewController.self)
-      .dependency { (scope, vc) in vc.service = *!scope }
+      .injection { $0.service = $1 }
     
-    builder.register(UIStoryboard.self)
+    builder.register(type: UIStoryboard.self)
       .lifetime(.single)
-      .initializer { DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: $0) }
+      .initial { DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: $0) }
     
     let container = try! builder.build()
-    let storyboard: UIStoryboard = *!container
+    let storyboard: UIStoryboard = try! *container
     
-    let viewController = storyboard.instantiateInitialViewController()
+		let navigation: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+		let viewController = navigation.childViewControllers[0]
     XCTAssert(viewController is TestViewController)
     guard let testVC = viewController as? TestViewController else {
       XCTFail("incorrect View Controller")
@@ -153,27 +156,28 @@ class DITranquillityTests_Storyboard: XCTestCase {
 	func test07_issue69_getViewControllerAfterDelay() {
 		let builder = DIContainerBuilder()
 		
-		builder.register(FooService.self)
-			.asType(ServiceProtocol.self)
-			.initializer { FooService() }
+		builder.register(type: FooService.self)
+			.as(ServiceProtocol.self).check{$0}
+      .initial{ FooService() }
 		
 		builder.register(vc: TestViewController.self)
-			.dependency { (scope, vc) in vc.service = *!scope }
+			.injection { scope, vc in try vc.service = *scope }
 		
-		builder.register(UIStoryboard.self)
+		builder.register(type: UIStoryboard.self)
 			.lifetime(.single)
-			.initializer { DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: $0) }
+			.initial { DIStoryboard(name: "TestStoryboard", bundle: Bundle(for: type(of: self)), container: $0) }
 		
 		var storyboard: UIStoryboard!
 		autoreleasepool {
 			let container = try! builder.build()
-			storyboard = *!container
+			storyboard = try! *container
 		}
 		
 		let semaphore = DispatchSemaphore(value: 0)
 		DispatchQueue.global(qos: .userInteractive).asyncAfter(wallDeadline: .now() + 1) {
 			defer { semaphore.signal() }
-			let viewController = storyboard.instantiateInitialViewController()
+			let navigation: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+			let viewController = navigation.childViewControllers[0]
 			XCTAssert(viewController is TestViewController)
 			guard let testVC = viewController as? TestViewController else {
 				XCTFail("incorrect View Controller")
@@ -183,7 +187,65 @@ class DITranquillityTests_Storyboard: XCTestCase {
 			XCTAssertEqual(testVC.service.foo(), "foo")
 		}
 		semaphore.wait()
-		
-		
 	}
+  
+  func test08_shortSyntaxInitialStoryboard() {
+    let builder = DIContainerBuilder()
+    
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial{ FooService() }
+    
+    builder.register(vc: TestViewController.self)
+      .injection { $0.service = $1 }
+    
+    builder.register(type: UIStoryboard.self)
+      .lifetime(.single)
+      .initial(name: "TestStoryboard", bundle: Bundle(for: type(of: self)))
+    
+    let container = try! builder.build()
+    let storyboard: UIStoryboard = try! *container
+    
+		let navigation: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+		let viewController = navigation.childViewControllers[0]
+    XCTAssert(viewController is TestViewController)
+    guard let testVC = viewController as? TestViewController else {
+      XCTFail("incorrect View Controller")
+      return
+    }
+    XCTAssertEqual(testVC.service.foo(), "foo")
+  }
+  
+  func test09_doubleVCfromStoryboard() {
+    let builder = DIContainerBuilder()
+    
+    builder.register(type: FooService.self)
+      .as(ServiceProtocol.self).check{$0}
+      .initial{ FooService() }
+    
+    builder.register(vc: TestViewController.self)
+      .injection { $0.service = $1 }
+      .lifetime(.lazySingle)
+    
+    builder.register(type: UIStoryboard.self)
+      .lifetime(.single)
+      .initial(name: "TestStoryboard", bundle: Bundle(for: type(of: self)))
+    
+    let container = try! builder.build()
+    let storyboard: UIStoryboard = try! *container
+    
+    let navigation1: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+    let viewController1 = navigation1.childViewControllers[0]
+    XCTAssert(viewController1 is TestViewController)
+    
+    let navigation2: UINavigationController = storyboard.instantiateInitialViewController() as! UINavigationController
+    let viewController2 = navigation2.childViewControllers[0]
+    XCTAssert(viewController2 is TestViewController)
+    
+    XCTAssert(viewController1 !== viewController2)
+    
+    let vc = viewController1 as! TestViewController
+    XCTAssertEqual(vc.service.foo(), "foo")
+    XCTAssertEqual(vc.service.foo(), "foo")
+  }
 }