+++ Samples/SampleHabr/SampleHabr/App/AppModule.swift
@@ -7,20 +7,13 @@
 //
 
 import DITranquillity
+import Logger
 
-class AppModule: DIModule {
-  func load(builder: DIContainerBuilder) {
-    builder.register(UIStoryboard.self)
-      .asName("Main")
-      .lifetime(.single)
-      .initializer { scope in DIStoryboard(name: "Main", bundle: nil, container: scope) }
-    
-		builder.register{ YourPresenter(server: *!$0) }
-      .lifetime(.perScope)
-      .dependency { (scope, self) in self.logger = *?scope }
-    
-		builder.register(vc: YourViewController.self)
-      .dependency { (scope, self) in self.presenter = try! scope.resolve() }
-  }
-
+public class AppModule: DIModule {
+  public var components: [DIComponent] { return [
+    ServerComponent(),
+    AppComponent()
+  ]}
+  
+  public var dependencies: [DIModule] { return [ LoggerModule() ] }
 }