+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# v2.0.0
+* Add documentation for migration
+* A full update documentation
+* Remove Lazy
+* Remove DynamicAssembly. Now there are late binding
+* Rename DIModule -> DIComponent and DIAssembly -> DIModule
+* Support area of vision for the components using modules
+* Add weak single
+* Add error description
+* Greatly improve old syntax
+* Separate library on modules: `Core`, `Description`, `Component`, `Module`, `Storyboard`, `Scan`, `RuntimeArgs`
+* Support auto inject properties for Obj-C types
+* Much more
+
+
+# v1.3.1
+* Support Lazy `DILazy`
+* Add documentation for lazy
+* Improved typing - now `DIType` it's `Any.Type` for a place `Any`
+
 # v1.3.0
 * Update documentation
 * Update README